fix cyclical import
