+++ packages/engine/src/avatar/components/AvatarAnimationComponent.ts
@@ -129,6 +129,7 @@ export const AvatarRigComponent = defineComponent({
     const visible = useOptionalComponent(entity, VisibleComponent)
     const modelComponent = useOptionalComponent(entity, ModelComponent)
     const locomotionAnimationState = useHookstate(getMutableState(AnimationState).loadedAnimations[locomotionAnimation])
+    const avatarComponent = useComponent(entity, AvatarComponent)
 
     useEffect(() => {
       if (!visible?.value || !debugEnabled.value || pending?.value || !rigComponent.value.normalizedRig?.hips?.node)
@@ -190,9 +191,37 @@ export const AvatarRigComponent = defineComponent({
       }
     }, [rigComponent.vrm])
 
+    useEffect(() => {
+      if (!rigComponent.normalizedRig.value) return
+      const rig = rigComponent.normalizedRig.value
+      rig.hips.node.getWorldPosition(hipsPos)
+      rig.head.node.getWorldPosition(headPos)
+      rig.leftFoot.node.getWorldPosition(leftFootPos)
+      rig.rightFoot.node.getWorldPosition(rightFootPos)
+      rig.leftLowerLeg.node.getWorldPosition(leftLowerLegPos)
+      rig.leftUpperLeg.node.getWorldPosition(leftUpperLegPos)
+      rig.leftEye ? rig.leftEye?.node.getWorldPosition(eyePos) : eyePos.copy(headPos)
+
+      avatarComponent.torsoLength.set(headPos.y - hipsPos.y)
+      avatarComponent.upperLegLength.set(hipsPos.y - leftLowerLegPos.y)
+      avatarComponent.lowerLegLength.set(leftFootPos.y - leftUpperLegPos.y)
+      avatarComponent.hipsHeight.set(hipsPos.y)
+      avatarComponent.eyeHeight.set(eyePos.y)
+      avatarComponent.footGap.set(footGap.subVectors(leftFootPos, rightFootPos).length())
+    }, [rigComponent.normalizedRig])
+
     return null
   }
 })
+
+const hipsPos = new Vector3(),
+  headPos = new Vector3(),
+  leftFootPos = new Vector3(),
+  rightFootPos = new Vector3(),
+  leftLowerLegPos = new Vector3(),
+  leftUpperLegPos = new Vector3(),
+  footGap = new Vector3(),
+  eyePos = new Vector3()
 
 /**Used to generate an offset map that retargets ik position animations to fit any rig */
 export const retargetIkUtility = (entity: Entity, bindTracks: KeyframeTrack[], height: number, flipped: boolean) => {
