+++ packages/engine/src/avatar/functions/spawnAvatarReceptor.ts
@@ -30,7 +30,7 @@ import { EntityUUID } from '@etherealengine/common/src/interfaces/EntityUUID'
 import { UserID } from '@etherealengine/engine/src/schemas/user/user.schema'
 import { getState } from '@etherealengine/hyperflux'
 
-import { setTargetCameraRotation } from '../../camera/systems/CameraInputSystem'
+import { setTargetCameraRotation } from '../../camera/functions/CameraFunctions'
 import { Engine } from '../../ecs/classes/Engine'
 import { Entity } from '../../ecs/classes/Entity'
 import { SceneState } from '../../ecs/classes/Scene'
