+++ packages/engine/src/camera/functions/CameraFunctions.ts
@@ -1,0 +1,80 @@
+import { ComponentType } from 'bitecs'
+import { clamp } from 'lodash'
+import { Entity } from '../../ecs/classes/Entity'
+import { getOptionalComponent, setComponent } from '../../ecs/functions/ComponentFunctions'
+import { FollowCameraComponent } from '../components/FollowCameraComponent'
+import { TargetCameraRotationComponent } from '../components/TargetCameraRotationComponent'
+
+export const setTargetCameraRotation = (entity: Entity, phi: number, theta: number, time = 0.3) => {
+  const cameraRotationTransition = getOptionalComponent(entity, TargetCameraRotationComponent) as
+    | ComponentType<typeof TargetCameraRotationComponent>
+    | undefined
+  if (!cameraRotationTransition) {
+    setComponent(entity, TargetCameraRotationComponent, {
+      phi: phi,
+      phiVelocity: { value: 0 },
+      theta: theta,
+      thetaVelocity: { value: 0 },
+      time: time
+    })
+  } else {
+    cameraRotationTransition.phi = phi
+    cameraRotationTransition.theta = theta
+    cameraRotationTransition.time = time
+  }
+}
+
+let lastScrollValue = 0
+
+/**
+ * Change camera distance.
+ * @param cameraEntity Entity holding camera and input component.
+ */
+export const handleCameraZoom = (cameraEntity: Entity, value: number): void => {
+  const scrollDelta = Math.sign(value - lastScrollValue) * 0.5
+  lastScrollValue = value
+
+  if (scrollDelta === 0) {
+    return
+  }
+
+  const followComponent = getOptionalComponent(cameraEntity, FollowCameraComponent) as
+    | ComponentType<typeof FollowCameraComponent>
+    | undefined
+
+  if (!followComponent) {
+    return
+  }
+
+  const epsilon = 0.001
+  const nextZoomLevel = clamp(followComponent.zoomLevel + scrollDelta, epsilon, followComponent.maxDistance)
+
+  // Move out of first person mode
+  if (followComponent.zoomLevel <= epsilon && scrollDelta > 0) {
+    followComponent.zoomLevel = followComponent.minDistance
+    return
+  }
+
+  // Move to first person mode
+  if (nextZoomLevel < followComponent.minDistance) {
+    followComponent.zoomLevel = epsilon
+    setTargetCameraRotation(cameraEntity, 0, followComponent.theta)
+    return
+  }
+
+  // Rotate camera to the top but let the player rotate if he/she desires
+  if (Math.abs(followComponent.maxDistance - nextZoomLevel) <= 1.0 && scrollDelta > 0) {
+    setTargetCameraRotation(cameraEntity, 85, followComponent.theta)
+  }
+
+  // Rotate from top
+  if (
+    Math.abs(followComponent.maxDistance - followComponent.zoomLevel) <= 1.0 &&
+    scrollDelta < 0 &&
+    followComponent.phi >= 80
+  ) {
+    setTargetCameraRotation(cameraEntity, 45, followComponent.theta)
+  }
+
+  followComponent.zoomLevel = nextZoomLevel
+}
