+++ packages/engine/src/camera/systems/CameraInputSystem.ts
@@ -23,7 +23,6 @@ All portions of the code written by the Ethereal Engine team are Copyright © 20
 Ethereal Engine. All Rights Reserved.
 */
 
-import { clamp } from 'lodash'
 import { Vector2 } from 'three'
 
 import { getState } from '@etherealengine/hyperflux'
@@ -36,13 +35,7 @@ import { throttle } from '../../common/functions/FunctionHelpers'
 import { Engine } from '../../ecs/classes/Engine'
 import { EngineState } from '../../ecs/classes/EngineState'
 import { Entity } from '../../ecs/classes/Entity'
-import {
-  ComponentType,
-  defineQuery,
-  getComponent,
-  getOptionalComponent,
-  setComponent
-} from '../../ecs/functions/ComponentFunctions'
+import { defineQuery, getComponent, getOptionalComponent } from '../../ecs/functions/ComponentFunctions'
 import { defineSystem } from '../../ecs/functions/SystemFunctions'
 import { InputSourceComponent } from '../../input/components/InputSourceComponent'
 
@@ -55,81 +48,8 @@ import { InputSystemGroup } from '../../ecs/functions/EngineFunctions'
 import { CameraSettings } from '../CameraState'
 import { FollowCameraComponent } from '../components/FollowCameraComponent'
 import { TargetCameraRotationComponent } from '../components/TargetCameraRotationComponent'
+import { handleCameraZoom, setTargetCameraRotation } from '../functions/CameraFunctions'
 import { CameraMode } from '../types/CameraMode'
-
-export const setTargetCameraRotation = (entity: Entity, phi: number, theta: number, time = 0.3) => {
-  const cameraRotationTransition = getOptionalComponent(entity, TargetCameraRotationComponent) as
-    | ComponentType<typeof TargetCameraRotationComponent>
-    | undefined
-  if (!cameraRotationTransition) {
-    setComponent(entity, TargetCameraRotationComponent, {
-      phi: phi,
-      phiVelocity: { value: 0 },
-      theta: theta,
-      thetaVelocity: { value: 0 },
-      time: time
-    })
-  } else {
-    cameraRotationTransition.phi = phi
-    cameraRotationTransition.theta = theta
-    cameraRotationTransition.time = time
-  }
-}
-
-let lastScrollValue = 0
-
-/**
- * Change camera distance.
- * @param cameraEntity Entity holding camera and input component.
- */
-export const handleCameraZoom = (cameraEntity: Entity, value: number): void => {
-  const scrollDelta = Math.sign(value - lastScrollValue) * 0.5
-  lastScrollValue = value
-
-  if (scrollDelta === 0) {
-    return
-  }
-
-  const followComponent = getOptionalComponent(cameraEntity, FollowCameraComponent) as
-    | ComponentType<typeof FollowCameraComponent>
-    | undefined
-
-  if (!followComponent) {
-    return
-  }
-
-  const epsilon = 0.001
-  const nextZoomLevel = clamp(followComponent.zoomLevel + scrollDelta, epsilon, followComponent.maxDistance)
-
-  // Move out of first person mode
-  if (followComponent.zoomLevel <= epsilon && scrollDelta > 0) {
-    followComponent.zoomLevel = followComponent.minDistance
-    return
-  }
-
-  // Move to first person mode
-  if (nextZoomLevel < followComponent.minDistance) {
-    followComponent.zoomLevel = epsilon
-    setTargetCameraRotation(cameraEntity, 0, followComponent.theta)
-    return
-  }
-
-  // Rotate camera to the top but let the player rotate if he/she desires
-  if (Math.abs(followComponent.maxDistance - nextZoomLevel) <= 1.0 && scrollDelta > 0) {
-    setTargetCameraRotation(cameraEntity, 85, followComponent.theta)
-  }
-
-  // Rotate from top
-  if (
-    Math.abs(followComponent.maxDistance - followComponent.zoomLevel) <= 1.0 &&
-    scrollDelta < 0 &&
-    followComponent.phi >= 80
-  ) {
-    setTargetCameraRotation(cameraEntity, 45, followComponent.theta)
-  }
-
-  followComponent.zoomLevel = nextZoomLevel
-}
 
 const avatarControllerQuery = defineQuery([AvatarControllerComponent])
